Fix Wayland support
