+++ cmake/linux/linux.cmake
@@ -1,4 +1,2 @@
-set(PLATFORM_SOURCES src/sys/linux/LinuxCap.cpp)
-set(PLATFORM_SOURCES src/sys/linux/desktopinfo.cpp)
-set(PLATFORM_SOURCES src/sys/linux/request.cpp)
+set(PLATFORM_SOURCES src/sys/linux/LinuxCap.cpp src/sys/linux/desktopinfo.cpp)
 set(PLATFORM_LIBRARIES dl)
